% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\docType{data}
\name{data}
\alias{data}
\title{Create and Customize Scatterplot}
\format{
An object of class \code{data.frame} with 812 rows and 30 columns.
}
\usage{
data
}
\arguments{
\item{data}{A dataframe containing the data.}
}
\value{
A customized scatterplot.
}
\description{
This function creates a scatterplot with customized aesthetics using ggplot2.
}
\keyword{datasets}
