% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_models.R
\name{knn_regressor}
\alias{knn_regressor}
\title{KNN Regressor}
\usage{
knn_regressor()
}
\value{
A list containing the R-squared value and Mean Squared Error (MSE) of the model.
}
\description{
Fits a K-Nearest Neighbors regression model using the provided dataset.
}
