% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_models.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear Regression}
\usage{
linear_regression()
}
\value{
A list containing the R-squared value and Mean Squared Error (MSE) of the model.
}
\description{
Fits a Linear Regression model using the provided dataset.
}
