% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_models.R
\name{random_forest_regressor}
\alias{random_forest_regressor}
\title{Random Forest Regressor}
\usage{
random_forest_regressor()
}
\value{
A list containing the R-squared value and Mean Squared Error (MSE) of the model.
}
\description{
Fits a Random Forest regression model using the provided dataset.
}
